requirements.txt
Flask==3.0.2
python-dotenv==1.0.1
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Cors==4.0.0
mysqlclient==2.2.4
PyMySQL==1.1.0
flask-restx==1.3.0 